cleanup: Make filesform.h and toxuri.cpp parseable by Linguist.
This fixes some warnings we get from the translation string updater.
